+++ tests/test_ortographic_camera.cpp
@@ -3,6 +3,6 @@
 
 TEST_CASE("it can init ortographic camera with default screen windows") {
     OrtographicCamera camera;
-    std::tuple<float, float, float, float> screenWindow = std::make_tuple(-1.555, 1.555, -1, 1);
+    std::tuple<double, double, double, double> screenWindow = std::make_tuple(-1.555, 1.555, -1, 1);
     CHECK_EQ(camera.getScreenWindow(), screenWindow);
 }
